requirements.txt
accelerate==0.25.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.0
aiosignal==1.4.0
annotated-types==0.7.0
async-timeout==5.0.1
attrs==25.4.0
cachetools==6.2.0
certifi==2025.10.5
charset-normalizer==3.4.3
colorama==0.4.6
datasets==2.16.1
dill==0.3.7
einops==0.8.1
filelock==3.19.1
frozenlist==1.8.0
fsspec==2023.10.0
google-ai-generativelanguage==0.6.15
google-api-core==2.26.0
google-api-python-client==2.184.0
google-auth==2.47.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.75.1
grpcio-status==1.71.2
httplib2==0.31.0
huggingface-hub==0.35.3
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
MarkupSafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
multiprocess==0.70.15
networkx==3.2.1
numpy==2.0.2
packaging==25.0
pandas==2.3.3
peft==0.7.1
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
psutil==7.1.0
pyarrow==21.0.0
pyarrow-hotfix==0.7
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.0
pydantic_core==2.41.1
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.3
regex==2025.9.18
requests==2.32.5
rsa==4.9.1
safetensors==0.6.2
six==1.17.0
sympy==1.14.0
tokenizers==0.15.2
torch==2.8.0
tqdm==4.67.1
transformers==4.36.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
protobuf==5.29.5
python-dotenv==1.2.2